Fixed bug in renameSvgFiles function and updated index.js file generation.
Previously, the renameSvgFiles function correctly renamed SVG files to JSX files, but did not add the newly renamed files to the list of JSX files to be included in the index.js file. This has been fixed by adding a line to add the newly renamed JSX file to the array of files to be included in index.js.

The updated function has been tested and is working correctly.